requirements.txt
aiohttp==3.9.3 ; python_version >= "3.10" and python_version < "3.12"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "3.12"
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "3.12"
anyio==4.2.0 ; python_version >= "3.10" and python_version < "3.12"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.11"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "3.12"
backoff==2.2.1 ; python_version >= "3.10" and python_version < "3.12"
beautifulsoup4==4.12.3 ; python_full_version >= "3.10.0" and python_version < "3.12"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "3.12"
chardet==5.2.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.12" and platform_system == "Windows"
crewai==0.11.0 ; python_version >= "3.10" and python_version < "3.12"
dataclasses-json==0.6.4 ; python_version >= "3.10" and python_version < "3.12"
deprecated==1.2.14 ; python_version >= "3.10" and python_version < "3.12"
distro==1.9.0 ; python_version >= "3.10" and python_version < "3.12"
docstring-parser==0.15 ; python_version >= "3.10" and python_version < "3.12"
emoji==2.10.1 ; python_full_version >= "3.10.0" and python_version < "3.12"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
filetype==1.2.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "3.12"
geojson==2.5.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
googleapis-common-protos==1.62.0 ; python_version >= "3.10" and python_version < "3.12"
greenlet==3.0.3 ; python_version >= "3.10" and python_version < "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.12"
httpcore==1.0.2 ; python_version >= "3.10" and python_version < "3.12"
httpx==0.26.0 ; python_version >= "3.10" and python_version < "3.12"
idna==3.6 ; python_version >= "3.10" and python_version < "3.12"
importlib-metadata==6.11.0 ; python_version >= "3.10" and python_version < "3.12"
instructor==0.5.2 ; python_version >= "3.10" and python_version < "3.12"
joblib==1.3.2 ; python_full_version >= "3.10.0" and python_version < "3.12"
jsonpatch==1.33 ; python_version >= "3.10" and python_version < "3.12"
jsonpointer==2.4 ; python_version >= "3.10" and python_version < "3.12"
langchain-community==0.0.20 ; python_version >= "3.10" and python_version < "3.12"
langchain-core==0.1.23 ; python_version >= "3.10" and python_version < "3.12"
langchain-openai==0.0.2.post1 ; python_version >= "3.10" and python_version < "3.12"
langchain==0.1.7 ; python_version >= "3.10" and python_version < "3.12"
langdetect==1.0.9 ; python_full_version >= "3.10.0" and python_version < "3.12"
langsmith==0.0.87 ; python_version >= "3.10" and python_version < "3.12"
lxml==5.1.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "3.12"
marshmallow==3.20.2 ; python_version >= "3.10" and python_version < "3.12"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "3.12"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.12"
nltk==3.8.1 ; python_full_version >= "3.10.0" and python_version < "3.12"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "3.12"
openai==1.12.0 ; python_version >= "3.10" and python_version < "3.12"
opentelemetry-api==1.22.0 ; python_version >= "3.10" and python_version < "3.12"
opentelemetry-exporter-otlp-proto-common==1.22.0 ; python_version >= "3.10" and python_version < "3.12"
opentelemetry-exporter-otlp-proto-http==1.22.0 ; python_version >= "3.10" and python_version < "3.12"
opentelemetry-proto==1.22.0 ; python_version >= "3.10" and python_version < "3.12"
opentelemetry-sdk==1.22.0 ; python_version >= "3.10" and python_version < "3.12"
opentelemetry-semantic-conventions==0.43b0 ; python_version >= "3.10" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.10" and python_version < "3.12"
protobuf==4.25.2 ; python_version >= "3.10" and python_version < "3.12"
pydantic-core==2.16.2 ; python_version >= "3.10" and python_version < "3.12"
pydantic==2.6.1 ; python_version >= "3.10" and python_version < "3.12"
pygments==2.17.2 ; python_version >= "3.10" and python_version < "3.12"
pyowm==3.3.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
pysocks==1.7.1 ; python_full_version >= "3.10.0" and python_version < "3.12"
python-dotenv==1.0.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
python-iso639==2024.2.7 ; python_full_version >= "3.10.0" and python_version < "3.12"
python-magic==0.4.27 ; python_full_version >= "3.10.0" and python_version < "3.12"
pytils==0.4.1 ; python_full_version >= "3.10.0" and python_version < "3.12"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.12"
rapidfuzz==3.6.1 ; python_full_version >= "3.10.0" and python_version < "3.12"
regex==2023.12.25 ; python_version >= "3.10" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.12"
requests[socks]==2.31.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
rich==13.7.0 ; python_version >= "3.10" and python_version < "3.12"
six==1.16.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "3.12"
soupsieve==2.5 ; python_full_version >= "3.10.0" and python_version < "3.12"
sqlalchemy==2.0.27 ; python_version >= "3.10" and python_version < "3.12"
tabulate==0.9.0 ; python_full_version >= "3.10.0" and python_version < "3.12"
tenacity==8.2.3 ; python_version >= "3.10" and python_version < "3.12"
tiktoken==0.5.2 ; python_version >= "3.10" and python_version < "3.12"
tools==0.1.9 ; python_full_version >= "3.10.0" and python_version < "3.12"
tqdm==4.66.2 ; python_version >= "3.10" and python_version < "3.12"
typer==0.9.0 ; python_version >= "3.10" and python_version < "3.12"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "3.12"
typing-inspect==0.9.0 ; python_version >= "3.10" and python_version < "3.12"
unstructured==0.10.25 ; python_full_version >= "3.10.0" and python_version < "3.12"
urllib3==2.2.0 ; python_version >= "3.10" and python_version < "3.12"
wrapt==1.16.0 ; python_version >= "3.10" and python_version < "3.12"
yarl==1.9.4 ; python_version >= "3.10" and python_version < "3.12"
zipp==3.17.0 ; python_version >= "3.10" and python_version < "3.12"
